Add render yaml and update project
