+++ Controllers/LeaderboardController.cs
@@ -59,100 +59,5 @@ namespace MyDotNetProject.Controllers
         }
     }
 
-    [ApiController]
-    [Route("api/[controller]")]
-    public class InvitationsController : ControllerBase
-    {
-        private readonly IInvitationService _invitationService;
 
-        public InvitationsController(IInvitationService invitationService)
-        {
-            _invitationService = invitationService;
-        }
-
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Invitation>> GetInvitation(string id)
-        {
-            var invitation = await _invitationService.GetInvitationByIdAsync(id);
-            if (invitation == null)
-            {
-                return NotFound();
-            }
-            return Ok(invitation);
-        }
-
-        [HttpGet("user/{userEmail}")]
-        public async Task<ActionResult<IEnumerable<Invitation>>> GetUserInvitations(string userEmail)
-        {
-            var invitations = await _invitationService.GetUserInvitationsAsync(userEmail);
-            return Ok(invitations);
-        }
-
-        [HttpGet("group/{groupId}")]
-        public async Task<ActionResult<IEnumerable<Invitation>>> GetGroupInvitations(string groupId)
-        {
-            var invitations = await _invitationService.GetGroupInvitationsAsync(groupId);
-            return Ok(invitations);
-        }
-
-        [HttpPost]
-        public async Task<ActionResult<Invitation>> CreateInvitation([FromBody] CreateInvitationRequest request)
-        {
-            try
-            {
-                var createdInvitation = await _invitationService.CreateInvitationAsync(
-                    request.GroupId, 
-                    request.UserEmail, 
-                    request.InvitedByUserId
-                );
-                return CreatedAtAction(nameof(GetInvitation), new { id = createdInvitation.Id }, createdInvitation);
-            }
-            catch (Exception ex)
-            {
-                return BadRequest(ex.Message);
-            }
-        }
-
-        [HttpPost("{id}/respond")]
-        public async Task<ActionResult> RespondToInvitation(string id, [FromBody] InvitationResponse response)
-        {
-            try
-            {
-                var result = await _invitationService.RespondToInvitationAsync(id, response.Accept, response.UserId);
-                if (!result)
-                {
-                    return BadRequest("Failed to respond to invitation");
-                }
-                return Ok();
-            }
-            catch (Exception ex)
-            {
-                return BadRequest(ex.Message);
-            }
-        }
-
-        [HttpDelete("{id}")]
-        public async Task<ActionResult> DeleteInvitation(string id)
-        {
-            var result = await _invitationService.DeleteInvitationAsync(id);
-            if (!result)
-            {
-                return NotFound();
-            }
-            return NoContent();
-        }
-    }
-
-    public class InvitationResponse
-    {
-        public bool Accept { get; set; }
-        public string UserId { get; set; } = string.Empty;
-    }
-
-    public class CreateInvitationRequest
-    {
-        public required string GroupId { get; set; }
-        public required string UserEmail { get; set; }
-        public required string InvitedByUserId { get; set; }
-    }
 }
