+++ appsettings.Production.json
@@ -4,8 +4,7 @@
     "ProductionConnection": "Host=YOUR_HOST;Database=YOUR_DB;Username=YOUR_USER;Password=YOUR_PASSWORD"
   },
   "MmaApi": {
-    "BaseUrl": "https://v1.mma.api-sports.io",
-    "ApiKey": "70657fde93dac699f8ab6e4a8cea01d9"
+    "BaseUrl": "https://v1.mma.api-sports.io"
   },
   "Firebase": {
     "ProjectId": "bet-app-c6402"
