+++ railway.toml
@@ -1,0 +1,21 @@
+# Railway deployment configuration
+# This file tells Railway how to deploy your .NET app
+
+$schema: https://railway.app/railway.schema.json
+
+build:
+  builder: DOCKERFILE
+  dockerfilePath: Dockerfile
+
+deploy:
+  startCommand: dotnet MyDotNetProject.dll
+  healthcheckPath: /api/users
+  healthcheckTimeout: 300
+  restartPolicyType: ON_FAILURE
+  restartPolicyMaxRetries: 10
+
+environments:
+  production:
+    variables:
+      ASPNETCORE_ENVIRONMENT: Production
+      ASPNETCORE_URLS: http://0.0.0.0:$PORT
