+++ temp_test.cs
@@ -1,0 +1,10 @@
+// Temporary file to trigger language server refresh
+using System;
+
+namespace Test
+{
+    public class TempTest
+    {
+        public string TestProperty { get; set; } = string.Empty;
+    }
+}
